Updated installment payment flow, user services, and added new validation for receipt payments

+++ app/Http/Requests/InstallmentPaymentRequest/StoreInstallmentPaymentData.php
@@ -29,7 +29,7 @@ class StoreInstallmentPaymentData extends FormRequest
     {
         return [
             'payment_date'=>'required|date',
-        'amount' => ['required', 'numeric', new StoreValidInstallmentAmount(Installment::findOrFail($this->route('id')))],
+        'amount' => ['required', 'integer', new StoreValidInstallmentAmount(Installment::findOrFail($this->route('id')))],
 
         ];
     }
